Refactore type generation to fix circular dependency
